Move MultiOrderedMap to utils package
